+++ packages/sample-lab/src/generation/errors.ts
@@ -1,0 +1,19 @@
+export class GenerationError extends Error {
+  readonly code: string;
+
+  constructor(code: string, message: string) {
+    super(message);
+    this.name = "GenerationError";
+    this.code = code;
+  }
+}
+
+export class CloudAndonError extends Error {
+  readonly code: string;
+
+  constructor(code: string, message: string) {
+    super(message);
+    this.name = "CloudAndonError";
+    this.code = code;
+  }
+}